chore: refactor the init script
