+++ app/internal/admin.py
@@ -1,6 +1,10 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends
+from ..dependencies import get_token_header
 
-router = APIRouter()
+router = APIRouter(
+    prefix='/admin',
+    tags=['admin'],
+    dependencies=[Depends(get_token_header)])
 
 
 @router.post('/')
